pyproject.toml
[project]
name = "faq-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Vinicius José",email = "vinijosenog@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"

dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langgraph (>=1.0.6,<2.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "langchain-neo4j (>=0.8.0,<0.9.0)",
    "neo4j-graphrag[sentence-transformers] (>=1.12.0,<2.0.0)",
    "langchain-groq (>=1.1.1,<2.0.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain (>=1.2.6,<2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (<5.0.0)",
    
]

[tool.poetry]
package-mode=false

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
pytest = "^9.0.2"
pytest-cov = ">=7.0.0,<8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]